# OpenAPI_Swagger.yaml
openapi: 3.0.3
info:
  title: Pet Store API
  description: API para la gestión de productos y servicios de una tienda de mascotas.
  version: 1.0.0
servers:
  - url: https://api.petstore.com/v1
...
paths:
  /api/products:
    get:
      tags:
        - products
      summary: Listar productos
      description: Obtiene una lista de productos.
      operationId: listProducts
      parameters:
        - name: category
          in: query
          description: Filtrar productos por categoría
          schema:
            type: string

      responses:
        '200':
          description: Lista de productos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Parámetros inválidos

    post:
      tags:
        - products
      summary: Crear un nuevo producto
      description: Permite a un administrador crear un nuevo producto.
      operationId: createProduct
      requestBody:
        description: Datos del nuevo producto
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Producto creado exitosamente
        '400':
          description: Datos de entrada inválidos
  /api/products/{productId}:
    parameters:
      - name: productId
        in: path
        description: ID del producto
        required: true
        schema:
          type: string

    get:
      tags:
        - products
      summary: Obtener detalles de un producto
      description: Obtiene la información de un producto específico.
      operationId: getProductById
      responses:
        '200':
          description: Detalles del producto obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Producto no encontrado

    put:
      tags:
        - products
      summary: Actualizar un producto
      description: Permite a un administrador actualizar los datos de un producto.
      operationId: updateProduct
      requestBody:
        description: Datos del producto a actualizar
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Producto actualizado correctamente
        '400':
          description: Datos de entrada inválidos
        '404':
          description: Producto no encontrado

    delete:
      tags:
        - products
      summary: Eliminar un producto
      description: Permite a un administrador eliminar un producto.
      operationId: deleteProduct
      responses:
        '200':
          description: Producto eliminado correctamente
        '404':
          description: Producto no encontrado
...
  /api/orders:
    post:
      tags:
        - orders
      summary: Crear un nuevo pedido
      description: Permite a un usuario crear un nuevo pedido.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Pedido creado exitosamente
        '400':
          description: Datos de entrada inválidos

    get:
      tags:
        - orders
      summary: Listar pedidos del usuario
      description: Devuelve la lista de pedidos de un usuario autenticado.
      operationId: listUserOrders
      responses:
        '200':
          description: Lista de pedidos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Usuario no autenticado
  /api/orders/{orderId}:
    parameters:
      - name: orderId
        in: path
        description: ID del pedido
        required: true
        schema:
          type: string

    get:
      tags:
        - orders
      summary: Obtener detalles de un pedido
      description: Permite a un usuario obtener los detalles de un pedido específico.
      operationId: getOrderDetails
      responses:
        '200':
          description: Detalles del pedido obtenidos exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Usuario no autenticado
        '404':
          description: Pedido no encontrado
  /api/orders/{orderId}/cancel:
    post:
      tags:
        - orders
      summary: Cancelar un pedido
      description: Permite a un usuario cancelar un pedido, si está permitido.
      operationId: cancelOrder
      parameters:
        - name: orderId
          in: path
          description: ID del pedido
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pedido cancelado exitosamente
        '400':
          description: No se puede cancelar el pedido
        '401':
          description: Usuario no autenticado
        '404':
          description: Pedido no encontrado

  # Endpoints de Reseñas y Valoraciones
...
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        description:
          type: string
        animalType:
          type: string
...
    CartItem:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
    Order:
      type: object
      properties:
        orderId:
          type: string
          description: ID único del pedido
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        status:
          type: string
          description: Estado actual del pedido
        totalAmount:
          type: number
          format: float
          description: Monto total del pedido
        createdDate:
          type: string
          format: date-time
          description: Fecha de creación del pedido
...